--- src/de/jost_net/JVerein/util/Spaltenauswahl.java
@@ -1,78 +1,0 @@
-/**********************************************************************
- * Copyright (c) by Heiner Jostkleigrewe
- * This program is free software: you can redistribute it and/or modify it under the terms of the 
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the 
- * License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,  but WITHOUT ANY WARRANTY; without 
- *  even the implied warranty of  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See 
- *  the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program.  If not, 
- * see <http://www.gnu.org/licenses/>.
- * 
- * heiner@jverein.de
- * www.jverein.de
- **********************************************************************/
-package de.jost_net.JVerein.util;
-
-import java.util.ArrayList;
-
-import de.jost_net.JVerein.gui.parts.JVereinTablePart;
-import de.jost_net.JVerein.rmi.Mitgliedstyp;
-import de.willuhn.jameica.gui.formatter.Formatter;
-import de.willuhn.jameica.system.Settings;
-
-public abstract class Spaltenauswahl
-{
-
-  private Settings settings;
-
-  private String tabelle;
-
-  private ArrayList<Spalte> spalten;
-
-  public Spaltenauswahl(String tabelle)
-  {
-    this.settings = new Settings(this.getClass());
-    this.tabelle = tabelle;
-    spalten = new ArrayList<>();
-  }
-
-  public void add(String spaltenbezeichnung, String spaltenname,
-      boolean defaultvalue, boolean auchNichtMitglied)
-  {
-    spalten.add(new Spalte(spaltenbezeichnung, spaltenname,
-        settings.getBoolean(tabelle + "." + spaltenname, defaultvalue),
-        auchNichtMitglied));
-  }
-
-  public void add(String spaltenbezeichnung, String spaltenname,
-      boolean defaultvalue, Formatter formatter, int align,
-      boolean auchNichtMitglied)
-  {
-    spalten.add(new Spalte(spaltenbezeichnung, spaltenname,
-        settings.getBoolean(tabelle + "." + spaltenname, defaultvalue),
-        formatter, align, auchNichtMitglied));
-  }
-
-  public void setColumns(JVereinTablePart part, int mitgliedstyp)
-  {
-    for (Spalte spalte : spalten)
-    {
-      if (mitgliedstyp == Integer.valueOf(Mitgliedstyp.MITGLIED)
-          || spalte.isAuchNichtMitglied())
-      {
-        part.addColumn(spalte.getSpaltenbezeichnung(), spalte.getSpaltenname(),
-            spalte.getFormatter(), false, spalte.getAlign(),
-            spalte.isChecked());
-      }
-    }
-  }
-
-  public ArrayList<Spalte> getSpalten()
-  {
-    return spalten;
-  }
-
-}
